add class npc
