+++ ai_mafia/llm/__init__.py
@@ -1,0 +1,1 @@
+from .llm_do import main
